+++ models/db/MotionSupporter.php
@@ -43,6 +43,67 @@ class MotionSupporter extends ISupporter
     }
 
     /**
+     * @return int[]
+     */
+    public static function getMyAnonymousSupportIds()
+    {
+        return \Yii::$app->session->get('anonymous_motion_supports', []);
+    }
+
+    /**
+     * @param MotionSupporter $support
+     */
+    public static function addAnonymouslySupportedMotion($support)
+    {
+        $pre   = \Yii::$app->session->get('anonymous_motion_supports', []);
+        $pre[] = IntVal($support->id);
+        \Yii::$app->session->set('anonymous_motion_supports', $pre);
+    }
+
+    /**
+     * @param Motion $motion
+     * @param User|null $user
+     * @param string $name
+     * @param string $orga
+     * @param int $role
+     */
+    public static function createSupport(Motion $motion, $user, $name, $orga, $role)
+    {
+        $maxPos = 0;
+        if ($user) {
+            foreach ($motion->motionSupporters as $supp) {
+                if ($supp->userId == $user->id) {
+                    $motion->unlink('motionSupporters', $supp, true);
+                } elseif ($supp->position > $maxPos) {
+                    $maxPos = $supp->position;
+                }
+            }
+        } else {
+            $alreadySupported = static::getMyAnonymousSupportIds();
+            foreach ($motion->motionSupporters as $supp) {
+                if (in_array($supp->id, $alreadySupported)) {
+                    $motion->unlink('motionSupporters', $supp, true);
+                } elseif ($supp->position > $maxPos) {
+                    $maxPos = $supp->position;
+                }
+            }
+        }
+
+        $support               = new MotionSupporter();
+        $support->motionId     = $motion->id;
+        $support->userId       = ($user ? $user->id : null);
+        $support->name         = $name;
+        $support->organization = $orga;
+        $support->position     = $maxPos + 1;
+        $support->role         = $role;
+        $support->save();
+
+        if (!$user) {
+            static::addAnonymouslySupportedMotion($support);
+        }
+    }
+
+    /**
      * @return array
      */
     public function rules()